disaster_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
requests==2.32.3
